+++ packages/plugins/Legend/CHANGELOG.md
@@ -1,8 +1,10 @@
 # CHANGELOG
 
-## unpublished
+## 1.1.2
 
 - Feature: Add option to configure used icons.
+- Chore: Upgrade peerDependency `@masterportal/masterportalapi` from `2.40.0` to `2.45.0`
+- Chore: Update `@polar/lib-custom-types` to `v2.0.0`.
 
 ## 1.1.1